requirements.txt
# Django Core
Django==5.2.8

# Database
# psycopg2-binary==2.9.9  # PostgreSQL uchun

# Image Processing
Pillow==10.1.0

# Utilities
python-slugify==8.0.1
python-dotenv==1.0.0

# Cache (Production uchun)
# django-redis==5.4.0
# redis==5.0.1

# Email
# django-anymail==10.2  # Email service providers uchun

# Security
# django-ratelimit==4.1.0  # Rate limiting

# Monitoring (Production uchun)
# sentry-sdk==1.39.1

# Development Tools
# django-debug-toolbar==4.2.0
# django-extensions==3.2.3

# Testing
# pytest==7.4.3
# pytest-django==4.7.0
# coverage==7.3.4

# Code Quality
# black==23.12.1
# flake8==7.0.0
# pylint==3.0.3

# API (agar kerak bo'lsa)
# djangorestframework==3.14.0
# drf-spectacular==0.27.0